fix switchup of user and profile pages
